Move hough.py module to new algorithms package
